Aufräumarbeiten, Testfunktionen ausgelagert Code kommentiert, nicht verwendeten entfernt Funktion zum umwandeln von Integerwerten in String ausgelagert Debugprint-Funktion ausgelagert
