+++ src/main/baro_driver.h
@@ -6,7 +6,7 @@
 uint8_t baro_init(void);
 uint8_t baro_read_int(uint8_t reg, int16_t *var);
 uint8_t baro_read_uint(uint8_t reg, uint16_t *var);
-uint8_t baro_raw_temp(int32_t *temp);
+uint8_t baro_temp_raw(int32_t *temp);
 uint8_t baro_temp(int32_t *temp);
 uint8_t baro_pres_raw(int32_t *pres, uint8_t oss);
 uint8_t baro_pres(int32_t *pres, uint8_t oss);
